+++ Rolling/Models/Rolls/DieRoll.cs
@@ -3,16 +3,16 @@ using Rolling.Utilities;
 
 namespace Rolling;
 
-public record DieRoll(int Result, int Size, Maybe<int> Id)
+public record DieRoll(int Result, int Size, Maybe<long> Id)
 {
     private static readonly Random s_random = new();
-    private static DieRoll One { get; } = new(1, 1, Maybe<int>.None);
+    private static DieRoll One { get; } = new(1, 1, Maybe<long>.None);
 
     public static DieRoll Random(int size)
     {
         if (size == 1)
             return One;
         
-        return new DieRoll(s_random.Next(1, size + 1), size, Maybe<int>.None);
+        return new DieRoll(s_random.Next(1, size + 1), size, Maybe<long>.None);
     }
 }
